+++ src/engine/TileMap/IsometricMap.ts
@@ -10,6 +10,9 @@ import { Entity } from '../EntityComponentSystem/Entity';
 import { DebugGraphicsComponent, ExcaliburGraphicsContext, Graphic, GraphicsComponent } from '../Graphics';
 import { IsometricEntityComponent } from './IsometricEntityComponent';
 import { DebugConfig } from '../Debug';
+import { PointerComponent } from '../Input/PointerComponent';
+import { PointerEvent } from '../Input/PointerEvent';
+
 export class IsometricTile extends Entity {
   /**
    * Indicates whether this tile is solid
@@ -304,6 +307,7 @@ export class IsometricMap extends Entity {
         type: CollisionType.Fixed
       }),
       new ColliderComponent(),
+      new PointerComponent(),
       new DebugGraphicsComponent((ctx, debugFlags) => this.debug(ctx, debugFlags), false)
     ], options.name);
     const { pos, tileWidth, tileHeight, columns: width, rows: height, renderFromTopOfGraphic, graphicsOffset, elevation } = options;
@@ -338,6 +342,33 @@ export class IsometricMap extends Entity {
         this.addChild(tile);
       }
     }
+
+    this._setupPointerToTile();
+  }
+
+  private _forwardPointerEventToTile = (eventType: string) => (evt: PointerEvent) => {
+    const tile = this.getTileByPoint(evt.worldPos);
+    if (tile) {
+      tile.events.emit(eventType, evt);
+    }
+  };
+
+  private _setupPointerToTile() {
+    this.events.on('pointerup', this._forwardPointerEventToTile('pointerup'));
+    this.events.on('pointerdown', this._forwardPointerEventToTile('pointerdown'));
+    this.events.on('pointermove', this._forwardPointerEventToTile('pointermove'));
+    this.events.on('pointerwheel', this._forwardPointerEventToTile('pointerwheel'));
+    this.events.on('pointercancel', this._forwardPointerEventToTile('pointercancel'));
+
+    // TODO do pointer enter,leave,drag even work like this? since we are forwarding events I don't think so
+    this.events.on('pointerenter', this._forwardPointerEventToTile('pointerenter'));
+    this.events.on('pointerleave', this._forwardPointerEventToTile('pointerleave'));
+
+    this.events.on('pointerdragstart', this._forwardPointerEventToTile('pointerdragstart'));
+    this.events.on('pointerdragend', this._forwardPointerEventToTile('pointerdragend'));
+    this.events.on('pointerdragenter', this._forwardPointerEventToTile('pointerdragenter'));
+    this.events.on('pointerdragleave', this._forwardPointerEventToTile('pointerdragleave'));
+    this.events.on('pointerdragmove', this._forwardPointerEventToTile('pointerdragmove'));
   }
 
   public update(): void {
@@ -386,6 +417,7 @@ export class IsometricMap extends Entity {
    * @param worldCoordinate
    */
   public worldToTile(worldCoordinate: Vector): Vector {
+    // TODO I don't think this handles parent transform see TileMap
     worldCoordinate = worldCoordinate.sub(this.transform.globalPos);
 
     const halfTileWidth = this.tileWidth / 2;